requirements.txt
Flask==1.0.2
google-cloud-storage==1.10.0
opencv-python==3.4.2.17
Pillow==5.2.0
requests==2.19.1
